Refactor calendars, introduce calendar customization
  - Split calendars into multiple files.
  - Add calendar customizatin mechanism (Islamic).
  - Update tests: fix estimated holidays names.
  - Update Pakistan Ashura dates.
