+++ holidays/countries/bahrain.py
@@ -9,9 +9,7 @@
 #  Website: https://github.com/dr-prodigy/python-holidays
 #  License: MIT (see LICENSE file)
 
-from datetime import date
-from datetime import timedelta as td
-
+from holidays.calendars import _CustomCalendar, _IslamicLunar
 from holidays.constants import FRI, SAT, MAY, JUL, AUG, OCT, DEC
 from holidays.holiday_base import HolidayBase
 from holidays.holiday_groups import InternationalHolidays, IslamicHolidays
@@ -31,7 +29,7 @@ class Bahrain(HolidayBase, InternationalHolidays, IslamicHolidays):
 
     def __init__(self, *args, **kwargs):
         InternationalHolidays.__init__(self)
-        IslamicHolidays.__init__(self)
+        IslamicHolidays.__init__(self, calendar=BahrainIslamicCalendar())
         super().__init__(*args, **kwargs)
 
     def _populate(self, year):
@@ -45,87 +43,27 @@ class Bahrain(HolidayBase, InternationalHolidays, IslamicHolidays):
 
         # Eid Al Fitr.
         eid_al_fitr = "Eid Al Fitr"
-        eid_al_fitr_mapping = {
-            2022: ((MAY, 2),),
-        }
-        if year in eid_al_fitr_mapping:
-            for dt in eid_al_fitr_mapping[year]:
-                eid_al_fitr_day = date(year, *dt)
-                self._add_holiday(eid_al_fitr, eid_al_fitr_day)
-                self._add_holiday(
-                    f"{eid_al_fitr} Holiday", eid_al_fitr_day + td(days=+1)
-                )
-                self._add_holiday(
-                    f"{eid_al_fitr} Holiday", eid_al_fitr_day + td(days=+2)
-                )
-        else:
-            self._add_eid_al_fitr_day(f"{eid_al_fitr}* (*estimated)")
-            self._add_eid_al_fitr_day_two(
-                f"{eid_al_fitr} Holiday* (*estimated)"
-            )
-            self._add_eid_al_fitr_day_three(
-                f"{eid_al_fitr} Holiday* (*estimated)"
-            )
+        self._add_eid_al_fitr_day(eid_al_fitr)
+        self._add_eid_al_fitr_day_two(f"{eid_al_fitr} Holiday")
+        self._add_eid_al_fitr_day_three(f"{eid_al_fitr} Holiday")
 
         # Eid Al Adha.
         eid_al_adha = "Eid Al Adha"
-        eid_al_adha_mapping = {
-            2022: ((JUL, 9),),
-        }
-        if year in eid_al_adha_mapping:
-            for dt in eid_al_adha_mapping[year]:
-                eid_al_adha_day = date(year, *dt)
-                self._add_holiday(eid_al_adha, eid_al_adha_day)
-                self._add_holiday(
-                    f"{eid_al_adha} Holiday", eid_al_adha_day + td(days=+1)
-                )
-                self._add_holiday(
-                    f"{eid_al_adha} Holiday", eid_al_adha_day + td(days=+2)
-                )
-        else:
-            self._add_eid_al_adha_day(f"{eid_al_adha}* (*estimated)")
-            self._add_eid_al_adha_day_two(f"{eid_al_adha}* (*estimated)")
-            self._add_eid_al_adha_day_three(f"{eid_al_adha}* (*estimated)")
+        self._add_eid_al_adha_day(eid_al_adha)
+        self._add_eid_al_adha_day_two(f"{eid_al_adha} Holiday")
+        self._add_eid_al_adha_day_three(f"{eid_al_adha} Holiday")
 
         # Al Hijra New Year.
         hijri_new_year = "Al Hijra New Year"
-        hijri_new_year_mapping = {
-            2022: ((JUL, 30),),
-        }
-        if year in hijri_new_year_mapping:
-            for dt in hijri_new_year_mapping[year]:
-                self._add_holiday(hijri_new_year, date(year, *dt))
-        else:
-            self._add_islamic_new_year_day(f"{hijri_new_year}* (*estimated)")
+        self._add_islamic_new_year_day(hijri_new_year)
 
         # Ashura.
         ashura = "Ashura"
-        ashura_mapping = {
-            2022: ((AUG, 7),),
-        }
-        if year in ashura_mapping:
-            for dt in ashura_mapping[year]:
-                ashura_day = date(year, *dt)
-                self._add_holiday(ashura, ashura_day)
-                self._add_holiday(
-                    f"{ashura} Holiday", ashura_day + td(days=+1)
-                )
-        else:
-            for ashura_dt in self._add_ashura_day(
-                f"{ashura} Holiday* (*estimated)"
-            ):
-                self._add_holiday(
-                    f"{ashura}* (*estimated)", ashura_dt + td(days=-1)
-                )
+        self._add_ashura_eve(f"{ashura} Holiday")
+        self._add_ashura_day(ashura)
 
         # Prophets Birthday.
-        mawlud = "Prophets Birthday"
-        mawlud_mapping = {2022: ((OCT, 8),)}
-        if year in mawlud_mapping:
-            for dt in mawlud_mapping[year]:
-                self._add_holiday(mawlud, date(year, *dt))
-        else:
-            self._add_mawlid_day(f"{mawlud}* (*estimated)")
+        self._add_mawlid_day("Prophets Birthday")
 
         # National Day.
         national_day = "National Day"
@@ -139,3 +77,25 @@ class BH(Bahrain):
 
 class BAH(Bahrain):
     pass
+
+
+class BahrainIslamicCalendar(_IslamicLunar, _CustomCalendar):
+    ASHURA_DATES = {
+        2022: ((AUG, 8),),
+    }
+
+    EID_AL_ADHA = {
+        2022: ((JUL, 9),),
+    }
+
+    EID_AL_FITR_DATES = {
+        2022: ((MAY, 2),),
+    }
+
+    HIJRI_NEW_YEAR_DATES = {
+        2022: ((JUL, 30),),
+    }
+
+    MAWLID_DATES = {
+        2022: ((OCT, 8),),
+    }
