+++ tests/countries/test_philippines.py
@@ -31,9 +31,9 @@ class TestPhilippines(TestCase):
             ("2022-04-15", "Good Friday"),
             ("2022-04-16", "Black Saturday"),
             ("2022-05-01", "Labour Day"),
-            ("2022-05-02", "Eid'l Fitr"),
+            ("2022-05-02", "Eid'l Fitr* (*estimated)"),
             ("2022-06-12", "Independence Day"),
-            ("2022-07-09", "Eid'l Adha"),
+            ("2022-07-09", "Eid'l Adha* (*estimated)"),
             ("2022-08-21", "Ninoy Aquino Day"),
             ("2022-08-29", "National Heroes Day"),
             ("2022-11-01", "All Saints' Day"),
