+++ tests/countries/test_uzbekistan.py
@@ -27,9 +27,9 @@ class TestUzbekistan(TestCase):
             ("2022-01-01", "New Year"),
             ("2022-03-08", "Women's Day"),
             ("2022-03-21", "Nauryz"),
-            ("2022-05-02", "Ramadan Khait"),
+            ("2022-05-02", "Ramadan Khait* (*estimated)"),
             ("2022-05-09", "Memorial Day"),
-            ("2022-07-09", "Kurban Khait"),
+            ("2022-07-09", "Kurban Khait* (*estimated)"),
             ("2022-09-01", "Independence Day"),
             ("2022-10-01", "Teacher's Day"),
             ("2022-12-08", "Constitution Day"),
